refactor: move util functions to utils.ts
